Type mismatch in fpclassify, isfinite and signbit without builtins.

+++ library/include/math.h
@@ -204,21 +204,21 @@ FP_NORMAL, FP_SUBNORMAL, FP_ZERO, x))
 #ifndef fpclassify
 extern int __fpclassify_float(float x);
 extern int __fpclassify_double(double x);
-extern int __fpclassify_long_double(double x);
+extern int __fpclassify_long_double(long double x);
 #define fpclassify(x) _Generic(x, float : __fpclassify_float, \
 double : __fpclassify_double, long double : __fpclassify_long_double)(x)
 #endif /* fpclassify */
 #ifndef isfinite
 extern int __isfinite_float(float x);
 extern int __isfinite_double(double x);
-extern int __isfinite_long_double(double x);
+extern int __isfinite_long_double(long double x);
 #define isfinite(x) _Generic(x, float : __isfinite_float, \
 double : __isfinite_double, long double : __isfinite_long_double)(x)
 #endif /* isfinite */
 #ifndef signbit
 extern int __signbit_float(float x);
 extern int __signbit_double(double x);
-extern int __signbit_long_double(double x);
+extern int __signbit_long_double(long double x);
 #define signbit(x) _Generic(x, float : __signbit_float, \
 double : __signbit_double, long double : __signbit_long_double)(x)
 #endif /* signbit */
